Update tests to reflect refactoring of session migration logic
